requirements.txt
chainer==7.1.0
cupy==7.1.0
cycler==0.10.0
decorator==4.4.1
fastrlock==0.4
filelock==3.0.12
imageio==2.6.1
imgaug==0.3.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.0.3
networkx==2.4
nltk==3.4.5
numpy==1.18.1
opencv-python-headless==4.1.2.30
Pillow==6.2.1
protobuf==3.7.1
pyparsing==2.4.5
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.15.0
scikit-learn==0.22
scipy==1.3.3
Shapely==1.6.4.post2
six==1.14.0
typing==3.6.6
typing-extensions==3.6.6
tensorboardX
